ejercicio  18 y 23 Ariana Orellana
