+++ src/ejercicio18-funciones.js
@@ -1,0 +1,4 @@
+export const CambiarTexto= () => {
+    const parrafo = document.getElementById('miParrafo');
+    parrafo.textContent='¡El Texto ha cambiado!';
+    };
