+++ src/ejercicio23-funciones.js
@@ -1,0 +1,5 @@
+export const actualizarSeleccion= (valor) => {
+    const result = document.getElementById('result');
+    result.innerHTML = `Has seleccionado: <strong>${valor}</strong>`;
+    console.log(`Lenguaje Seleccionado: ${valor}`);
+};
